backend/requirements.txt
django>=5.0,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.0,<5.0
python-dotenv>=1.0,<2.0
requests>=2.31,<3.0
